reimpl(rdCamera): BuildFOV + BuildClipFrustum
rdCamera_BuildClipFrustum: stub -> full (plane-normal vector math).

rdCamera_BuildFOV was transcribed from a corrupted pre-ftol decompile;
the clean decompile exposes three defects, all verified against the
disassembly at 0x48fdc0:
- ortho branch wrote the wrong clip-frustum fields (shifted 8 bytes):
  now writes nearPlane/farPlane/orthoLeftPlane/orthoTopPlane and zeroes
  the correct set.
- perspective orthoRightPlane was inverted (aspect/(h/fovY) vs the real
  (h/fovY)/aspect).
- camera->unk was truncated via (int); the binary stores a float
  (depth-range normalization scale 1/(fovY/zFar - fovY/zNear)), so the
  field is retyped uint32_t -> float.

+++ src/types.h
@@ -259,7 +259,7 @@ extern "C"
         void (*fnProject)(rdVector3*, rdVector3*); // 0x4c
         void (*fnProjectLst)(rdVector3*, rdVector3*, unsigned int); // 0x50
         float ambientLight;
-        uint32_t unk;
+        float unk; // 1 / (fov_y/zFar - fov_y/zNear); depth-range normalization scale
         rdVector4 unk2;
         int numLights; // 0x6c
         rdLight* lights[128]; // Jkdf 64, Indy 128. really 128 ?
